+++ src/main/java/com/fabiogouw/bank/adapters/actors/messages/BalanceResponse.java
@@ -1,15 +1,16 @@
 package com.fabiogouw.bank.adapters.actors.messages;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class BalanceResponse implements Serializable {
     private static final long serialVersionUID = 1302757287444314441L;
-    private final double _balance;
+    private final BigDecimal _balance;
 
-    public BalanceResponse(double balance) {
+    public BalanceResponse(BigDecimal balance) {
         _balance = balance;
     }
-    public double getBalance() {
+    public BigDecimal getBalance() {
         return _balance;
     }
 }
