+++ src/main/java/com/fabiogouw/bank/adapters/actors/messages/OperationRequest.java
@@ -1,11 +1,13 @@
 package com.fabiogouw.bank.adapters.actors.messages;
 
+import java.math.BigDecimal;
+
 public class OperationRequest extends AccountMessage {
     private static final long serialVersionUID = -188612147356070992L;
     private final String _correlationId;
-    private final double _amount;
+    private final BigDecimal _amount;
     
-    public OperationRequest(String accountId, String correlationId, double amount) {
+    public OperationRequest(String accountId, String correlationId, BigDecimal amount) {
         super(accountId);
         _correlationId = correlationId;
         _amount = amount;
@@ -13,7 +15,7 @@ public class OperationRequest extends AccountMessage {
     public String getCorrelationId() {
         return _correlationId;
     }
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return _amount;
     }
 }
