+++ src/main/java/com/fabiogouw/bank/adapters/dtos/BalanceReservation.java
@@ -2,12 +2,14 @@ package com.fabiogouw.bank.adapters.dtos;
 
 import akka.actor.ActorRef;
 
+import java.math.BigDecimal;
+
 public class BalanceReservation {
     private String _correlationId;
-    private double _amount;
+    private BigDecimal _amount;
     private ActorRef _originalRequester;
 
-    public BalanceReservation(String correlationId, double amount, ActorRef originalRequester) {
+    public BalanceReservation(String correlationId, BigDecimal amount, ActorRef originalRequester) {
         _correlationId = correlationId;
         _amount = amount;
         _originalRequester = originalRequester;
@@ -16,7 +18,7 @@ public class BalanceReservation {
     public String getCorrelationId() {
         return _correlationId;
     }
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return _amount;
     }
     public ActorRef getOriginalRequester() {
