+++ src/main/java/com/fabiogouw/bank/core/contracts/Ledger.java
@@ -1,5 +1,6 @@
 package com.fabiogouw.bank.core.contracts;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
@@ -8,6 +9,6 @@ import com.fabiogouw.bank.core.domain.Account;
 
 public interface Ledger {
     public CompletableFuture<Double> getBalance(String accountId);
-    public CompletableFuture<Void> insert(String accountId, Date entryDatetime, UUID entryId, double amount, UUID correlationId, String description, int entryType);
-    public CompletableFuture<Void> saveBalance(String accountId, Date snapshotDate, double balance);
+    public CompletableFuture<Void> insert(String accountId, Date entryDatetime, UUID entryId, BigDecimal amount, UUID correlationId, String description, int entryType);
+    public CompletableFuture<Void> saveBalance(String accountId, Date snapshotDate, BigDecimal balance);
 }
